+++ src/pages/MyPage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { useAuthStore } from '../stores/useAuthStore';
 import { AuthAPI, UserAPI, SpecAPI } from '../api';
 import { getAccessToken } from '../utils/token';
-import ProfileSection from '../components/user/ProfileSection';
+import ProfileSection from '../components/user/ProfileCard/ProfileSection';
 import MenuList from '../components/user/MenuList';
 import WithdrawModal from '../components/user/WithdrawModal';
